refactor: extract pipeline preset registry to rag_pipeline_presets.py (#364)
Stage 1 of the rag_core.py decomposition epic (external senior review
2026-05 finding #3, follow-up to PR-D #357). After PR-D the file
still weighed ~4,154 lines; this PR shaves ~80 net by lifting the
pipeline preset *registry* — pure static data plus three
self-contained lookup helpers — into a brand-new leaf module
rag_pipeline_presets.py (140 lines).

Moved to the new module:
- Defaults: DEFAULT_CLI_PIPELINE_NAME, DEFAULT_RAG_PIPELINE_NAME.
- RRF fusion knob + range: RRF_K = 60, VALID_RRF_K_RANGE.
- PIPELINE_CONFIG_KEYS (the override-able tuple).
- DEFAULT_COMPARISON_BALANCE (consumed by the preset entries).
- PIPELINE_PRESETS (naive_baseline / agentic_full / agentic_full_llm —
  ADR 0001 + ADR 0011 contracts preserved bit-for-bit).
- PIPELINE_ALIASES (full / full_llm).
- Helpers: is_pipeline_name, pipeline_cli_choices,
  canonical_pipeline_name.

rag_core.py re-exports all 11 symbols via `from rag_pipeline_presets
import (...)` so the five known external consumers (app.py,
api/main.py, scripts/run_benchmark.py, demo/streamlit_app.py,
eval/run_eval.py) and the test suite keep importing them from
rag_core unchanged.

resolve_pipeline_config intentionally stays in rag_core for stage 1 —
it depends on VALID_RETRIEVAL_MODES, VALID_RETRIEVAL_BACKENDS, and
VALID_BM25_STOPWORD_PROFILES, all retrieval-side validation that is
itself a candidate for stage 2. The new module remains a pure leaf —
no imports from rag_core, no circular dependency risk.

QUERY_TYPE_TOP_K_DEFAULTS and MAX_AGENT_ITERATIONS also stay — they
belong with the retrieval / agent loop machinery, not with the preset
registry.

pyproject.toml adds rag_pipeline_presets to the coverage source list
so the new module participates in the PR #325 measurement loop.

Verification (ADR 0001 invariant + 5b escape clause):
- Pre-extraction parity: 8 module-level constants + 3 helpers each
  type-checked and == against the live rag_core values before the
  inline definitions were deleted.
- Post-extraction re-export identity: rag_core.X is
  rag_pipeline_presets.X holds for every moved symbol.
- resolve_pipeline_config still resolves naive_baseline,
  agentic_full, agentic_full_llm, and the full alias correctly, and
  still raises ValueError on invalid names.
- Full pytest (EMBEDDING_BACKEND=hashing): 659 passed, 0 failed,
  121 subtests passed. No behavior change in retrieval / verifier /
  answer path.

Closes #364
